+++ src/components/Icon/icons/IconAttention/IconAttenion.styles.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/css';
+
+import { colors } from '../../../../variables';
+
+export const styles = css`
+  width: 12px;
+  height: 12px;
+  fill: ${colors.dark65};
+`;