+++ src/components/Icon/icons/IconBoldArrowRight/IconBoldArrowRight.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './icon-bold-arrow-right.svg';
+
+import { styles } from './IconBoldArrowRight.styles';
+
+export const IconBoldArrowRight = ({ className, ...props }: GenericIconProps) => (
+  <Icon {...props} className={cx(styles, className)} glyph={image} />
+);