+++ src/components/Icon/icons/IconCalendar/IconCalendar.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/css';
+
+export const styles = css`
+  width: 16px;
+  height: 16px;
+  fill: rgba(0, 0, 0, 0.65);
+`;