+++ src/components/Icon/icons/IconClock/IconClock.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './clock.svg';
+
+import { styles } from './IconClock.styles';
+
+export const IconClock = ({ className, ...props }: GenericIconProps) => (
+  <Icon {...props} className={cx(styles, className)} glyph={image} />
+);