+++ src/components/Icon/icons/IconCluster/IconCluster.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/css';
+
+export const styles = css`
+  width: 14px;
+  height: 14px;
+  fill: #fff;
+`;