+++ src/components/Icon/icons/IconEdit/IconEdit.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './edit.svg';
+
+import { styles } from './IconEdit.styles';
+
+export const IconEdit = ({ className, onClick }: GenericIconProps) => (
+  <Icon className={cx(styles, className)} glyph={image} onClick={onClick} />
+);