+++ src/components/Icon/icons/IconFile/IconFile.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './file.svg';
+
+import { styles } from './IconFile.styles';
+
+export const IconFile = ({ className, onClick }: GenericIconProps) => (
+  <Icon className={cx(styles, className)} glyph={image} onClick={onClick} />
+);