+++ src/components/Icon/icons/IconGeoPin/IconGeoPin.styles.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/css';
+
+import { colors } from '../../../../variables';
+
+export const styles = css`
+  width: 14px;
+  height: 14px;
+  fill: ${colors.intentDanger};
+`;