+++ src/components/Icon/icons/IconOk/IconOk.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './ok.svg';
+
+import { styles } from './IconOk.styles';
+
+export const IconOk = ({ className, ...props }: GenericIconProps) => (
+  <Icon {...props} className={cx(styles, className)} glyph={image} />
+);