+++ src/components/Icon/icons/IconPlay/IconPlay.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { cx } from '@emotion/css';
+
+import type { GenericIconProps } from '../../Icon';
+import { Icon } from '../../Icon';
+import image from './play.svg';
+
+import { styles } from './IconPlay.styles';
+
+export const IconPlay = ({ className, ...props }: GenericIconProps) => (
+  <Icon {...props} className={cx(styles, className)} glyph={image} />
+);