+++ src/components/Icon/icons/IconSearch/IconSearch.styles.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/css';
+
+import { colors } from '../../../../variables';
+
+export const styles = css`
+  width: 16px;
+  height: 16px;
+  fill: ${colors.dark25};
+`;